+++ results/2plus1D/user_mesh_spm_1plus1D.py
@@ -5,7 +5,7 @@ import sys
 # set logging level
 pybamm.set_logging_level("INFO")
 
-# load (1+1D) SPMe model
+# load (1+1D) SPM model
 options = {
     "current collector": "potential pair",
     "dimensionality": 1,
@@ -35,7 +35,7 @@ param.process_model(model)
 param.process_geometry(geometry)
 
 # set mesh using user-supplied edges in z
-z_edges = np.array([0, 0.03, 0.1, 0.3, 0.47, 0.5, 0.73, 0.8, 0.911, 1])
+z_edges = np.array([0, 0.025, 0.05, 0.1, 0.3, 0.5, 0.7, 0.9, 0.95, 0.975, 1])
 submesh_types = model.default_submesh_types
 submesh_types["current collector"] = pybamm.MeshGenerator(
     pybamm.UserSupplied1DSubMesh, submesh_params={"edges": z_edges}